Add auth check to intermediate api(DB communication)

+++ apis/intermediate_api/intermediate_api/api/intermediate_api.py
@@ -21,20 +21,26 @@ def get_auth0_user_key_info(auth):
     as we want to keep the users_permissions table up to date
     for the dashboards.
     """
-    app.logger.info(
-        f"Received request at {const.INTERMEDIATE_API_AUTH0_USER_INFO_ENDPOINT}"
+    if auth:
+        app.logger.info(
+            f"Received request at {const.INTERMEDIATE_API_AUTH0_USER_INFO_ENDPOINT}"
+        )
+
+        token = auth0.get_token_auth_header()
+        unverified_claims = jwt.get_unverified_claims(token)
+
+        user_id = unverified_claims["sub"].split("|")[1]
+        permission_list = unverified_claims["permissions"]
+
+        # Update the Users_Permissions table
+        db.update_user_access_permission(user_id, permission_list)
+
+        return jsonify({"permissions": permission_list, "id": user_id})
+
+    return (
+        jsonify({"Warning": "You do not have permission to access."}),
+        const.UNAUTHORIZED_CODE,
     )
-
-    token = auth0.get_token_auth_header()
-    unverified_claims = jwt.get_unverified_claims(token)
-
-    user_id = unverified_claims["sub"].split("|")[1]
-    permission_list = unverified_claims["permissions"]
-
-    # Update the Users_Permissions table
-    db.update_user_access_permission(user_id, permission_list)
-
-    return jsonify({"permissions": permission_list, "id": user_id})
 
 
 # Edit User
@@ -48,11 +54,17 @@ def get_auth0_users(auth):
     setting user permission to users that don't
     exist in the DB yet
     """
-    app.logger.info(
-        f"Received request at {const.INTERMEDIATE_API_AUTH0_USERS_ENDPOINT}"
+    if auth:
+        app.logger.info(
+            f"Received request at {const.INTERMEDIATE_API_AUTH0_USERS_ENDPOINT}"
+        )
+        response, status_code = auth0.get_users()
+        return jsonify(response), status_code
+
+    return (
+        jsonify({"Warning": "You do not have permission to access."}),
+        const.UNAUTHORIZED_CODE,
     )
-    response, status_code = auth0.get_users()
-    return jsonify(response), status_code
 
 
 @app.route(const.INTERMEDIATE_API_ALL_PRIVATE_PROJECTS_ENDPOINT, methods=["GET"])
@@ -60,10 +72,16 @@ def get_auth0_users(auth):
 @decorators.endpoint_exception_handler
 def get_private_projects(auth):
     """Fetching all private projects from the Project table"""
-    app.logger.info(
-        f"Received request at {const.INTERMEDIATE_API_ALL_PRIVATE_PROJECTS_ENDPOINT}"
+    if auth:
+        app.logger.info(
+            f"Received request at {const.INTERMEDIATE_API_ALL_PRIVATE_PROJECTS_ENDPOINT}"
+        )
+        return jsonify(db.get_projects("private"))
+
+    return (
+        jsonify({"Warning": "You do not have permission to access."}),
+        const.UNAUTHORIZED_CODE,
     )
-    return jsonify(db.get_projects("private"))
 
 
 @app.route(const.INTERMEDIATE_API_ALL_PUBLIC_PROJECTS_ENDPOINT, methods=["GET"])
@@ -83,13 +101,19 @@ def get_user_allowed_projects(auth):
     """Fetching all the projects that are already allowed to a user
     Will be used for allowed Private Projects dropdown
     """
-    app.logger.info(
-        f"Received request at {const.INTERMEDIATE_API_USER_PROJECTS_ENDPOINT}"
+    if auth:
+        app.logger.info(
+            f"Received request at {const.INTERMEDIATE_API_USER_PROJECTS_ENDPOINT}"
+        )
+
+        user_id = request.args.to_dict()["user_id"]
+
+        return jsonify(db.get_user_project_permission(user_id))
+
+    return (
+        jsonify({"Warning": "You do not have permission to access."}),
+        const.UNAUTHORIZED_CODE,
     )
-
-    user_id = request.args.to_dict()["user_id"]
-
-    return jsonify(db.get_user_project_permission(user_id))
 
 
 @app.route(const.INTERMEDIATE_API_USER_ALLOCATE_PROJECTS_ENDPOINT, methods=["POST"])
@@ -97,18 +121,24 @@ def get_user_allowed_projects(auth):
 @decorators.endpoint_exception_handler
 def allocate_projects_to_user(auth):
     """Allocate the chosen project(s) to the chosen user."""
-    app.logger.info(
-        f"Received request at {const.INTERMEDIATE_API_USER_ALLOCATE_PROJECTS_ENDPOINT}"
+    if auth:
+        app.logger.info(
+            f"Received request at {const.INTERMEDIATE_API_USER_ALLOCATE_PROJECTS_ENDPOINT}"
+        )
+
+        data = json.loads(request.data.decode())
+
+        user_id = data["user_info"]["value"]
+        project_list = data["project_info"]
+
+        db.allocate_projects_to_user(user_id, project_list)
+
+        return Response(status=const.OK_CODE)
+
+    return (
+        jsonify({"Warning": "You do not have permission to access."}),
+        const.UNAUTHORIZED_CODE,
     )
-
-    data = json.loads(request.data.decode())
-
-    user_id = data["user_info"]["value"]
-    project_list = data["project_info"]
-
-    db.allocate_projects_to_user(user_id, project_list)
-
-    return Response(status=const.OK_CODE)
 
 
 @app.route(const.INTERMEDIATE_API_USER_REMOVE_PROJECTS_ENDPOINT, methods=["POST"])
@@ -116,18 +146,24 @@ def allocate_projects_to_user(auth):
 @decorators.endpoint_exception_handler
 def remove_projects_from_user(auth):
     """Remove the chosen project(s) from the chosen user."""
-    app.logger.info(
-        f"Received request at {const.INTERMEDIATE_API_USER_REMOVE_PROJECTS_ENDPOINT}"
+    if auth:
+        app.logger.info(
+            f"Received request at {const.INTERMEDIATE_API_USER_REMOVE_PROJECTS_ENDPOINT}"
+        )
+
+        data = json.loads(request.data.decode())
+
+        user_id = data["user_info"]["value"]
+        project_list = data["project_info"]
+
+        db.remove_projects_from_user(user_id, project_list)
+
+        return Response(status=const.OK_CODE)
+
+    return (
+        jsonify({"Warning": "You do not have permission to access."}),
+        const.UNAUTHORIZED_CODE,
     )
-
-    data = json.loads(request.data.decode())
-
-    user_id = data["user_info"]["value"]
-    project_list = data["project_info"]
-
-    db.remove_projects_from_user(user_id, project_list)
-
-    return Response(status=const.OK_CODE)
 
 
 @app.route(const.INTERMEDIATE_API_ALL_USERS_PROJECTS_ENDPOINT, methods=["GET"])
@@ -135,11 +171,17 @@ def remove_projects_from_user(auth):
 @decorators.endpoint_exception_handler
 def get_all_users_projects(auth):
     """Pull every assigned project for all users from Users_Projects table"""
-    app.logger.info(
-        f"Received request at {const.INTERMEDIATE_API_ALL_USERS_PROJECTS_ENDPOINT}"
+    if auth:
+        app.logger.info(
+            f"Received request at {const.INTERMEDIATE_API_ALL_USERS_PROJECTS_ENDPOINT}"
+        )
+        auth0_users, _ = auth0.get_users()
+        return db.get_all_users_project_permissions(auth0_users)
+
+    return (
+        jsonify({"Warning": "You do not have permission to access."}),
+        const.UNAUTHORIZED_CODE,
     )
-    auth0_users, _ = auth0.get_users()
-    return db.get_all_users_project_permissions(auth0_users)
 
 
 @app.route(const.INTERMEDIATE_API_ALL_PERMISSIONS_ENDPOINT, methods=["GET"])
@@ -147,10 +189,16 @@ def get_all_users_projects(auth):
 @decorators.endpoint_exception_handler
 def get_all_permissions(auth):
     """Pull all possible access permission (Auth0_Permission table)"""
-    app.logger.info(
-        f"Received request at {const.INTERMEDIATE_API_ALL_PERMISSIONS_ENDPOINT}"
+    if auth:
+        app.logger.info(
+            f"Received request at {const.INTERMEDIATE_API_ALL_PERMISSIONS_ENDPOINT}"
+        )
+        return jsonify({"all_permissions": db.get_all_permissions_for_dashboard()})
+
+    return (
+        jsonify({"Warning": "You do not have permission to access."}),
+        const.UNAUTHORIZED_CODE,
     )
-    return jsonify({"all_permissions": db.get_all_permissions_for_dashboard()})
 
 
 @app.route(const.INTERMEDIATE_API_ALL_USERS_PERMISSIONS_ENDPOINT, methods=["GET"])
@@ -158,11 +206,17 @@ def get_all_permissions(auth):
 @decorators.endpoint_exception_handler
 def get_all_users_permissions(auth):
     """Pull every assigned access permission for all uesrs from Users_Permissions table"""
-    app.logger.info(
-        f"Received request at {const.INTERMEDIATE_API_ALL_USERS_PERMISSIONS_ENDPOINT}"
+    if auth:
+        app.logger.info(
+            f"Received request at {const.INTERMEDIATE_API_ALL_USERS_PERMISSIONS_ENDPOINT}"
+        )
+        auth0_users, _ = auth0.get_users()
+        return db.get_all_users_permissions(auth0_users)
+
+    return (
+        jsonify({"Warning": "You do not have permission to access."}),
+        const.UNAUTHORIZED_CODE,
     )
-    auth0_users, _ = auth0.get_users()
-    return db.get_all_users_permissions(auth0_users)
 
 
 @app.route(const.INTERMEDIATE_API_CREATE_PROJECT_ENDPOINT, methods=["POST"])
@@ -170,39 +224,45 @@ def get_all_users_permissions(auth):
 @decorators.endpoint_exception_handler
 def create_project(auth):
     """Create new project(s)"""
-    app.logger.info(
-        f"Received request at {const.INTERMEDIATE_API_CREATE_PROJECT_ENDPOINT}"
-    )
-    data = json.loads(request.data.decode())
-
-    # Use Display Name as Project ID
-    # If Project ID is not provided
-    if data.get("id") is None or len(data["id"].strip()) == 0:
-        data["id"] = data["name"].lower().strip().replace(" ", "_")
-
-    if db.is_project_in_db(data["id"]):
-        return (
-            jsonify({"project_created": False, "reason": "Already exists"}),
-            const.CONFLICT_CODE,
+    if auth:
+        app.logger.info(
+            f"Received request at {const.INTERMEDIATE_API_CREATE_PROJECT_ENDPOINT}"
         )
-    else:
-        response_code = utils.proxy_to_core_api(
-            request,
-            const.PROJECT_CREATE_NEW_ENDPOINT,
-            "POST",
-            PROJECT_API_BASE,
-            PROJECT_API_TOKEN,
-            auth,
-            data=json.dumps(data),
-            user_id=auth0.get_user_id(),
-            action="Project Creation - Triggering project generation",
-        ).status_code
+        data = json.loads(request.data.decode())
 
-        if response_code == const.OK_CODE:
-            db.add_project_to_db(data["id"], data["name"], data["access_level"])
-            return jsonify({"project_created": True}), const.OK_CODE
-        else:
+        # Use Display Name as Project ID
+        # If Project ID is not provided
+        if data.get("id") is None or len(data["id"].strip()) == 0:
+            data["id"] = data["name"].lower().strip().replace(" ", "_")
+
+        if db.is_project_in_db(data["id"]):
             return (
-                jsonify({"project_created": False}),
-                response_code,
+                jsonify({"project_created": False, "reason": "Already exists"}),
+                const.CONFLICT_CODE,
             )
+        else:
+            response_code = utils.proxy_to_core_api(
+                request,
+                const.PROJECT_CREATE_NEW_ENDPOINT,
+                "POST",
+                PROJECT_API_BASE,
+                PROJECT_API_TOKEN,
+                auth,
+                data=json.dumps(data),
+                user_id=auth0.get_user_id(),
+                action="Project Creation - Triggering project generation",
+            ).status_code
+
+            if response_code == const.OK_CODE:
+                db.add_project_to_db(data["id"], data["name"], data["access_level"])
+                return jsonify({"project_created": True}), const.OK_CODE
+            else:
+                return (
+                    jsonify({"project_created": False}),
+                    response_code,
+                )
+
+    return (
+        jsonify({"Warning": "You do not have permission to access."}),
+        const.UNAUTHORIZED_CODE,
+    )
